pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datascience-agent"
version = "0.5.1"
description = "AI Agent with dynamic planning and persistent Jupyter kernel execution for data analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "DSAgent Contributors" }
]
keywords = [
    "ai",
    "agent",
    "llm",
    "jupyter",
    "data-analysis",
    "machine-learning",
    "planner",
    "dsagent",
    "datascience-agent",
    "autonomous-agent"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    # Core
    "litellm>=1.0.0",
    "jupyter-client>=8.0.0",
    "ipykernel>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "watchdog>=3.0.0",
    # CLI
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "pyyaml>=6.0.0",
    # Data Science
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "pycaret>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sse-starlette>=1.0.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.scripts]
dsagent = "dsagent.cli.main:main"

[project.urls]
Homepage = "https://github.com/nmlemus/dsagent"
Documentation = "https://github.com/nmlemus/dsagent#readme"
Repository = "https://github.com/nmlemus/dsagent"

[tool.hatch.build.targets.wheel]
packages = ["src/dsagent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
